pyproject.toml
[project]
name = "meet-your-partner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "google-cloud-bigquery>=3.36.0",
]
